maua-esports/src/horas-pae-relatorios/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
pandas==1.5.3
fpdf2==2.7.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0